refactor: AP부분 리팩토링
