+++ src/index.js
@@ -6,7 +6,7 @@ import App from './App';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import ScrollToTop from '@articles/ScrollToTop';
-import CheckWallet from '@components/wallets/components/CheckWallet';
+import CheckWallet from '@articles/CheckWallet';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
